requirements.txt
streamlit==1.36.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-image==0.24.0
pillow==10.4.0
pytesseract==0.3.10
yfinance==0.2.40
ta==0.11.0
matplotlib==3.9.0